File lifecycle update
Added a file manager that tracks a history of changes and allows them to be
undone and redone. The application now warns you when you have unsaved changes
when you exit, load, or start a new file. Added save-as functionality. Also
added an icon, updated menus, and fixed some tests.