Improve .app startup time by switching pyinstaller from onefile to onedir
I was under the impression that because MacOS applications are distributed as single .app files, you have to use the pyinstaller `onefile` option when building them. Because of the large size of the application's dependencies, opening the it would take forever as they are all unzipped. It turns out that there is no such requirement, and that if you are willing to accept a larger application, you can get much faster startup times by using `onedir`. The application is still a single .app file, but it opens much faster because the dependencies aren't zipped into a single file internally.